FIX: home page pagination not working
When the books were not sorted by a category, pagination failed. The culprit was the endpoint mapping. Endpoint mappings are greedy, so having the {category}/{page} pattern before the Books/{page} was causing the Books/1 url to be captured by {category}/{page}. assigning category = "Books", which doesn't exist. Doing so caused the LINQ Where expression to fail, and thus return no results.

Moved the Books/{page} pattern above the {category}/{page} pattern to force better matching.